+++ include/tao/pegtl/contrib/parse_tree.hpp
@@ -21,6 +21,7 @@
 #include "../apply_mode.hpp"
 #include "../config.hpp"
 #include "../demangle.hpp"
+#include "../eol.hpp"
 #include "../memory_input.hpp"
 #include "../normal.hpp"
 #include "../nothing.hpp"
@@ -105,7 +106,7 @@ namespace TAO_PEGTL_NAMESPACE::parse_tree
          return { m_begin.data, m_end.data };
       }
 
-      template< tracking_mode P = tracking_mode::eager, typename Eol = eol::lf_crlf >
+      template< tracking_mode P = tracking_mode::eager, typename Eol = ascii::lf_crlf >
       [[nodiscard]] memory_input< P, Eol > as_memory_input() const
       {
          assert( has_content() );