+++ include/tao/pegtl/mmap_input.hpp
@@ -18,7 +18,7 @@
 
 namespace TAO_PEGTL_NAMESPACE
 {
-   template< tracking_mode P = tracking_mode::eager, typename Eol = eol::lf_crlf >
+   template< tracking_mode P = tracking_mode::eager, typename Eol = ascii::lf_crlf >
    struct mmap_input
       : private internal::mmap_file,
         public memory_input< P, Eol >