+++ src/test/pegtl/restart_input.cpp
@@ -11,7 +11,7 @@ namespace TAO_PEGTL_NAMESPACE
    void test_lazy()
    {
       const std::string data = "abc";
-      memory_input< tracking_mode::lazy, eol::lf_crlf, std::string > in( data, __FUNCTION__ );
+      memory_input< tracking_mode::lazy, ascii::lf_crlf, std::string > in( data, __FUNCTION__ );
       bool success = parse< grammar >( in );
       TAO_PEGTL_TEST_ASSERT( success );
       in.restart();
@@ -22,7 +22,7 @@ namespace TAO_PEGTL_NAMESPACE
    void test_eager()
    {
       const std::string data = "abc";
-      memory_input< tracking_mode::eager, eol::lf_crlf, std::string > in( std::string_view{ data }, __FUNCTION__ );
+      memory_input< tracking_mode::eager, ascii::lf_crlf, std::string > in( std::string_view{ data }, __FUNCTION__ );
       bool success = parse< grammar >( in );
       TAO_PEGTL_TEST_ASSERT( success );
       in.restart();