+++ src/test/pegtl/verify_rule.hpp
@@ -8,7 +8,6 @@
 #include <cstdlib>
 #include <string>
 
-#include <tao/pegtl/eol.hpp>
 #include <tao/pegtl/memory_input.hpp>
 #include <tao/pegtl/tracking_mode.hpp>
 #include <tao/pegtl/type_list.hpp>
@@ -34,7 +33,7 @@ namespace TAO_PEGTL_NAMESPACE
       {}
    };
 
-   template< typename Rule, typename Eol = eol::lf_crlf >
+   template< typename Rule, typename Eol = ascii::lf_crlf >
    void verify_rule( const std::size_t line, const char* file, const std::string& data, const result_type expected, int remain = -1 )
    {
       if( remain < 0 ) {
@@ -60,7 +59,7 @@ namespace TAO_PEGTL_NAMESPACE
       }
    }
 
-   template< typename Rule, typename Eol = eol::lf_crlf >
+   template< typename Rule, typename Eol = ascii::lf_crlf >
    void verify_only( const std::size_t line, const char* file, const std::string& data, const result_type expected, const std::size_t remain )
    {
       {